Clean branch of DFP pack version 1.0.0-rc5
Updates for 1.0.0-rc5 version of pack:
    - Support for Event Router
    - Support for I2S DMA
    - Support for LP peripherals : LPSPI, LPUART, LPPDM, LPI2S, LPGPIO
    - Support for CMSIS OSPI Driver & ospi xip code
    - Support for CMSIS Ethernet MAC/Phy Drivers with LWIP stack &
      http server application
    - Updated various Baremetal & FreeRTOS test apps for B0 Flat board.
    - Updated CPI and MIPI CSI drivers (Validated on COB board)
    - Add missing interrupts in startup files
    - Correct M55 version
    - Add alias for local peripheral(MHU, DMA, NPU) interrupts
    - Improvement in RTSS Cache Management
    - Tune M55 Prefetch control for performance

Known Issues:
    - I3C is working only with Group B pins (BBS-84)
    - MIPI CSI is not working on FLAT board.(BECP-1424)

+++ Alif_CMSIS/Source/DPHY_Private.h
@@ -78,9 +78,9 @@ typedef enum _DPHY_MODE_CFG {
 
 /** \brief hsfreqrange and osc_freq_target range */
 typedef struct _DPHY_FREQ_RANGE {
-	uint32_t bitrate_in_mbps;           /**< DPHY data rate in mbps            */
+	uint16_t bitrate_in_mbps;           /**< DPHY data rate in mbps            */
 	uint8_t  hsfreqrange;               /**< DPHY HS frequency range           */
-	uint32_t osc_freq_target;           /**< DPHY oscillator frequency target  */
+	uint16_t osc_freq_target;           /**< DPHY oscillator frequency target  */
 }DPHY_FREQ_RANGE;
 
 /** \brief PLL vco_cntrl range */
